a "small" refactor and fixing
