+++ src/tables/mod.rs
@@ -1,3 +1,2 @@
-
-pub mod zobrist;
 pub mod piece_sq_tables;
+pub mod zobrist;
