+++ src/tables/zobrist.rs
@@ -1,8 +1,8 @@
-use std::{collections::HashMap, fmt::Debug};
+use chess::{Board, ChessMove, Color, Piece, Square};
 use nohash_hasher::BuildNoHashHasher;
-use std::collections::VecDeque;
-use chess::{Piece, Board, Square, Color, ChessMove};
 use rand::Rng;
+use std::collections::VecDeque;
+use std::{collections::HashMap, fmt::Debug};
 
 #[derive(Debug, Clone)]
 pub struct NodeInfo {
@@ -24,10 +24,13 @@ fn bytes_to_capacity<K, V>(total_bytes: usize) -> usize {
     total_bytes / (std::mem::size_of::<K>() + std::mem::size_of::<V>())
 }
 
-impl<V> ZobristHashMap<V> where V: Debug {
+impl<V> ZobristHashMap<V>
+where
+    V: Debug,
+{
     pub fn new(byte_size: usize) -> Self {
         let capacity = bytes_to_capacity::<u64, V>(byte_size);
-        ZobristHashMap { 
+        ZobristHashMap {
             map: HashMap::with_hasher(BuildNoHashHasher::default()),
             zobrist_table: init_zobrist(),
             keys: VecDeque::with_capacity(capacity),
@@ -68,7 +71,8 @@ pub fn init_zobrist() -> ([[u64; 6]; 64], [[u64; 6]; 64]) {
     let mut rng = rand::thread_rng();
     let mut black_zobrist_table: [[u64; 6]; 64] = [[0; 6]; 64];
     let mut white_zobrist_table: [[u64; 6]; 64] = [[0; 6]; 64];
-    for sq in 0..64 { // 0 to 63 (inclusive)
+    for sq in 0..64 {
+        // 0 to 63 (inclusive)
         for piece_index in Piece::Pawn.to_index()..Piece::King.to_index() {
             white_zobrist_table[sq][piece_index] = rng.gen_range(0..u64::MAX);
             black_zobrist_table[sq][piece_index] = rng.gen_range(0..u64::MAX);
@@ -78,19 +82,26 @@ pub fn init_zobrist() -> ([[u64; 6]; 64], [[u64; 6]; 64]) {
     (white_zobrist_table, black_zobrist_table)
 }
 
-pub fn hash(board: &Board, white_zobrist_table: [[u64; 6]; 64], black_zobrist_table: [[u64; 6]; 64]) -> u64 {
+pub fn hash(
+    board: &Board,
+    white_zobrist_table: [[u64; 6]; 64],
+    black_zobrist_table: [[u64; 6]; 64],
+) -> u64 {
     let mut final_hash = 0;
     for i in 0..64 {
         if let Some(piece) = board.piece_on(unsafe { Square::new(i) }) {
             if let Some(color) = board.color_on(unsafe { Square::new(i) }) {
                 match color {
-                    Color::White => final_hash = final_hash ^ white_zobrist_table[i as usize][piece.to_index()],
-                    Color::Black => final_hash = final_hash ^ black_zobrist_table[i as usize][piece.to_index()],
+                    Color::White => {
+                        final_hash = final_hash ^ white_zobrist_table[i as usize][piece.to_index()]
+                    }
+                    Color::Black => {
+                        final_hash = final_hash ^ black_zobrist_table[i as usize][piece.to_index()]
+                    }
                 };
             }
         }
     }
-    
+
     final_hash
 }
-
